Fitsmart/requirements.txt
streamlit
mysql-connector-python
bcrypt
fpdf
openai
python-dotenv
plotly
PyMuPDF